proof-of-concept implementation
This is a simple POC of the build-observer concept that delegates the
actual eBPF work to bpftrace.

Future work will move to C + github.com/cilium/ebpf to simplify the
setup and requirements for end users.

There is also an existing implementation based on ptrace that might be
useful (but much less performant) on platforms without eBPF support.